feat: permitir existencias decimales

+++ src/main/java/com/kynsoft/report/applications/command/almacenproducto/entrada/EntradaAlmacenCommand.java
@@ -17,7 +17,7 @@ public class EntradaAlmacenCommand implements ICommand {
     private String descripcion;
     private String numeroFactura;
     private String centroCosto;
-    private Integer stockNuevo;
+    private Double stockNuevo;
 
     public EntradaAlmacenCommand(UUID almacenFincaProductoId, Integer cantidad,
                                   TipoMovimientoStock tipo, String descripcion,
