+++ src/main/java/com/kynsoft/report/applications/command/almacenproducto/entrada/EntradaAlmacenMessage.java
@@ -11,6 +11,6 @@ import java.util.UUID;
 public class EntradaAlmacenMessage implements ICommandMessage {
     private UUID almacenFincaProductoId;
     private Integer cantidad;
-    private Integer stockNuevo;
+    private Double stockNuevo;
     private final String command = "ENTRADA_ALMACEN";
 }
