+++ src/main/java/com/kynsoft/report/applications/command/fincaproducto/asignar/AsignarProductoAFincaRequest.java
@@ -10,6 +10,6 @@ import java.util.UUID;
 public class AsignarProductoAFincaRequest {
     private UUID fincaId;
     private UUID productoId;
-    private Integer stock;
-    private Integer stockMinimo;
+    private Double stock;
+    private Double stockMinimo;
 }
