+++ src/main/java/com/kynsoft/report/applications/command/fincaproducto/entradafactura/EntradaFacturaCommandHandler.java
@@ -16,7 +16,7 @@ public class EntradaFacturaCommandHandler implements ICommandHandler<EntradaFact
     public void handle(EntradaFacturaCommand command) {
         // Get current stock before operation
         FincaProductoDto dto = service.getById(command.getId());
-        Integer stockAnterior = dto.getStock();
+        Double stockAnterior = dto.getStock();
 
         // Perform the entry
         service.entradaFactura(
