+++ src/main/java/com/kynsoft/report/domain/dto/BalanceProductoDto.java
@@ -14,9 +14,9 @@ public class BalanceProductoDto {
     private UUID fincaProductoId;
     private String productoCode;
     private String productoName;
-    private Integer stockInicial;
-    private Long totalEntradas;
-    private Long totalSalidas;
-    private Integer stockFinal;
+    private Double stockInicial;
+    private Double totalEntradas;
+    private Double totalSalidas;
+    private Double stockFinal;
     private Long cantidadMovimientos;
 }
